# trayex-backend/openapi.yaml
openapi: 3.0.3
info:
  title: Trayex API (MVP)
  version: 0.1.0
servers:
  - url: http://localhost:4000
paths:
...
  /tickets/{reservationId}/qr:
    get:
      summary: Obtener QR offline firmado
      parameters:
        - in: path
          name: reservationId
          required: true
          schema: { type: string }
      responses: { '200': { description: OK } }
  /validate:
    post:
      summary: Validar QR (online, demo)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                qr: { type: string }
      responses: { '200': { description: OK } }
...